requirements.txt
# Real-Time Alertness Monitor - Dependencies

# Core ML frameworks
torch>=2.0.0
torchvision>=0.15.0

# Computer vision
opencv-python>=4.8.0
ultralytics>=8.0.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0

# ML utilities
scikit-learn>=1.3.0

# Transformers
transformers>=4.30.0

# Data download
kagglehub>=0.1.0

# Audio
playsound>=1.3.0

# Image processing
Pillow>=10.0.0
